requirements.txt
python-dotenv  # for loading the local environmental file
slack_bolt  # for connecting the app to slack
